feat(auth): initialized database state for our auth
