+++ app/models/user.server.ts
@@ -1,4 +1,4 @@
-import type { Password, User } from '@prisma/client'
+import type { User } from '@prisma/client'
 import bcrypt from 'bcryptjs'
 
 import { prisma } from '~/db.server'
@@ -19,11 +19,7 @@ export async function createUser(email: User['email'], password: string) {
 	return prisma.user.create({
 		data: {
 			email,
-			password: {
-				create: {
-					hash: hashedPassword,
-				},
-			},
+			password: hashedPassword,
 		},
 	})
 }
@@ -34,20 +30,17 @@ export async function deleteUserByEmail(email: User['email']) {
 
 export async function verifyLogin(
 	email: User['email'],
-	password: Password['hash']
+	password: User['password']
 ) {
 	const userWithPassword = await prisma.user.findUnique({
 		where: { email },
-		include: {
-			password: true,
-		},
 	})
 
 	if (!userWithPassword || !userWithPassword.password) {
 		return null
 	}
 
-	const isValid = await bcrypt.compare(password, userWithPassword.password.hash)
+	const isValid = await bcrypt.compare(password, userWithPassword.password)
 
 	if (!isValid) {
 		return null
