+++ app/root.tsx
@@ -22,7 +22,7 @@ export const links: LinksFunction = () => {
 
 export const meta: MetaFunction = () => ({
 	charset: 'utf-8',
-	title: 'Remix Notes',
+	title: 'SimpleBudget',
 	viewport: 'width=device-width,initial-scale=1',
 })
