+++ app/routes/login.tsx
@@ -77,7 +77,7 @@ export const meta: MetaFunction = () => {
 
 export default function LoginPage() {
 	const [searchParams] = useSearchParams()
-	const redirectTo = searchParams.get('redirectTo') || '/notes'
+	const redirectTo = searchParams.get('redirectTo') || '/'
 	const actionData = useActionData() as ActionData
 	const emailRef = React.useRef<HTMLInputElement>(null)
 	const passwordRef = React.useRef<HTMLInputElement>(null)
