--- app/routes/notes/index.tsx
@@ -1,12 +1,0 @@
-import { Link } from '@remix-run/react'
-
-export default function NoteIndexPage() {
-	return (
-		<p>
-			No note selected. Select a note on the left, or{' '}
-			<Link to="new" className="text-blue-500 underline">
-				create a new note.
-			</Link>
-		</p>
-	)
-}
